backend/requirements.txt
requests==2.28.1
django==3.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.9.3
django-filter==21.1
python-dotenv==0.20.0
djoser==2.1.0
drf-extra-fields==3.4.0
pillow==9.2.0
django-colorfield
webcolors==1.11.1